player and asteroids

+++ circleshape.py
@@ -3,7 +3,7 @@ import pygame
 
 
 class CircleShape(pygame.sprite.Sprite):
-    def __init__(self, x: int, y: int, radius: int) -> None:
+    def __init__(self, x: float, y: float, radius: int) -> None:
         if hasattr(self, "containers"):
             super().__init__(self.containers)
         else:
